hw-2 update: BaseAbstractAddressModel, City, and Country classes are placed into core/models and price fields replaced with DecimalField
