+++ ecommerce/customers/models.py
@@ -8,7 +8,7 @@ from django.utils.translation import gettext_lazy as _
 
 from django.db import models
 
-from core.models import BaseAbstractModel
+from core.models import BaseAbstractModel, BaseAbstractAddressModel
 from customers.managers import CustomerManager
 
 class Customer(AbstractBaseUser, PermissionsMixin):
@@ -69,47 +69,9 @@ class Customer(AbstractBaseUser, PermissionsMixin):
         """Send an email to this user."""
         send_mail(subject, message, from_email, [self.email], **kwargs)
 
-class Country(BaseAbstractModel):
+
+class Address(BaseAbstractAddressModel):
     name = models.CharField(max_length=255, verbose_name=_("Name"))
-
-    class Meta:
-        verbose_name = _("country")
-        verbose_name_plural = _("countries")
-
-    def __str__(self):
-        return f"{self.name}"
-
-
-class City(BaseAbstractModel):
-    name = models.CharField(max_length=255, verbose_name=_("Name"))
-    country = models.ForeignKey(Country, on_delete=models.CASCADE)
-
-    class Meta:
-        verbose_name = _("city")
-        verbose_name_plural = _("cities")
-
-    def __str__(self):
-        return f"{self.name}"
-
-
-class Address(BaseAbstractModel):
-    name = models.CharField(max_length=255, verbose_name=_("Name"))
-    full_name = models.CharField(max_length=255, verbose_name=_("Full Name"))
-    line1 = models.CharField(max_length=255, verbose_name=_("Line 1"))
-    line2 = models.CharField(max_length=255, verbose_name=_("Line 2"))
-    # phone = RegexValidator(r'^(5)([0-9]{2})\s?([0-9]{3})\s?([0-9]{2})\s?([0-9]{2})$', message="Phone number must be entered in the format: '555 987 65 43' or '5559876543'.") # TR specific
-    phone_validator = RegexValidator(regex=r'^\+?1?\d{9,15}$',
-                           message="Phone number must be entered in the format: '+999999999'. Up to 15 digits allowed.")
-    phone = models.CharField(max_length=255, verbose_name=_("Phone"), validators=[phone_validator])
-
-    district = models.CharField(max_length=255, verbose_name=_("District"))
-
-    postal_code_validator = RegexValidator(regex=r'(?i)^[a-z0-9][a-z0-9\- ]{0,10}[a-z0-9]$',
-                                 message="Invalid format.")  # https://stackoverflow.com/a/19844362/12579069
-    postal_code = models.CharField(max_length=255, verbose_name=_("Postal Code"), validators=[postal_code_validator])
-
-    city = models.ForeignKey(City, on_delete=models.CASCADE)
-
     customer = models.ForeignKey(Customer, on_delete=models.CASCADE)
 
     class Meta:
